An S3 issue fix.
Fixed issue #215.
If the user chooses EU buckets, bucketnames that are not valid hostname labels will be rejected.
If the user does not choose EU buckets, a warning is displayed for bucketnames that are not valid hostname labels, and the setup is switched to the path-style calling method.

Fixed issue #214.
There were multiple issues with entering a bucket/folder type path, those should now be resolved.

Fixed issue #212.
The UI and commandline now supports Reduced Redundancy Storage.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@430 59da171f-624f-0410-aa54-27559c288bec