Update issue #252.
Now also cleans the backup manifest.
This bug means that the after a delete-older-than or delete-all-but-n-full would leave the backup manifest.
Issuing the same command again would then remove the manifests.
This should not be noticeable by regular users, except that the lingering manifests may be displayed when attempting a restore operation.


git-svn-id: https://duplicati.googlecode.com/svn/trunk@477 59da171f-624f-0410-aa54-27559c288bec