Update issue #90. Status: Fixed
I have fixed a number of UI issues when using async uploads.
For one, the pause/stop features now work.

If the progress is paused, a upload progressbar will be displayed,
so the progress does not appear to hang.

When the backup is complete, the remaining uploads
are now done as normal, resulting in a normal UI display with progress.

I have also reserved some space on the progress bar for cleanup,
and inserted a status message indicating that the cleanup is in progress.

The async now works so well, that I think it should be on by default.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@482 59da171f-624f-0410-aa54-27559c288bec