Update issue #174 Status: Fixed The new strategy is to provide general assembly redirects. These are generated during the build procedure and added to the output. Using these redirects it is possible to load a new/old version of a backend (or other module) into Duplicati.
To ease testing, it is now also possible to place a dll for a backend in
a subfolder named "backends" and it will override the normal backend.
To remove the test backend, simply stop Duplicati and delete the "backends" folder.

The same works for compression, encryption, generic modules, and settingcontrols 
with the folders "compression", "encryption", "modules", and "settingcontrols" 
respectively.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@536 59da171f-624f-0410-aa54-27559c288bec