Update issue #285 Status: Fixed
I found out that Mono on windows does not support "mixed-mode" assemblies, which is what the System.Data.SQLite.dll's are:
http://www.mono-project.com/CPlusPlus

On my machine it gives a weird error with "CRT not initialized" if I run Duplicati under Mono.
I have changed the loader to check for Linux AND Mono.
I have also update the build script to include the version of the dll that you found,
it seems that Mono only runs with 32bit on windows, so the dll must also be 32 bit.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@562 59da171f-624f-0410-aa54-27559c288bec