Update issue #275 Status: Fixed The logic was attempting to prevent a case where the backup was scheduled to run and "Run backup now" was checked. In this special case, the "Run backup now" does nothing because the backup will run twice otherwise. The logic was a bit flawed, so based on some various settings it would detect that it was scheduled to run, while it was not.
Thanks to your fine description, I was able to reproduce easily and fix it.
With the regards to the other error message, I have reproduced it by doing the following without restarting:
1) create the backup
2) edit the backup
3) delete the backup

I will create a new issue for it, but the bug is deep in the database abstraction layer,
so it is pretty hard to debug.

Update issue #227
Status: Fixed
Now correctly removes the deleted task from the scheduler and updates the display.
I also managed to get a slight speedup on the display update.



git-svn-id: https://duplicati.googlecode.com/svn/trunk@572 59da171f-624f-0410-aa54-27559c288bec