Update issue #187
I have updated the app.config file to include the useLegacyV2RuntimeActivationPolicy="true" attribute, and I can open and run it directly from VS2010 (after upgrade wizard),
and it seems to retain the attribute, and not cause trouble for 2.0.

If someone wants to produce a installer from a VS2010 build, beware that the build script uses a custom app.config file which will overwrite this one.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@575 59da171f-624f-0410-aa54-27559c288bec