Update issue #317 Status: Fixed
Turns out it was a more generic problem than I first assumed.
The problem was that there are two similar controls on the page (the duration editors),
and their internal controls are named the same, so it applied the setting from the top control to the lower,
and the top control was set to "each day".

The other problem "system is slow" is also fixed, although I'm unsure why it happened in the first place.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@630 59da171f-624f-0410-aa54-27559c288bec