Update issue #337 Status: Fixed
It was supposed to write an error message like:
When listing the folder <foldername> the file <filename> was listed, but the server now reports that the file is not found.
This can be because the file is deleted or unavailable, but it can also be because the file extension .aes is blocked by the web server. IIS blocks unknown extensions by default.
Error message: 404 - File not found

Unfortunately the error message was incorrectly formated so you got a highly misleading message instead.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@658 59da171f-624f-0410-aa54-27559c288bec