This commit upgrades the manifest version number to 3. This commit includes an option to create an unencrypted verification file. There is now a check to see if all volumes mentioned in the manifest exists, before performing backup/restore operations. This check is also performed on cleanup. There is now a check that prevents auto-cleanup from removing more than 2 orphan files.
If a delete-older-than or delete-all-but-n-full is interrupted, the leftover
files will not be auto-deleted but must be deleted with the cleanup command,
which allows the user to see what files are delete.

Update issue #360
Status: Fixed

There is now a complete verification chain, where the manifest hashes are verified backwards.


git-svn-id: https://duplicati.googlecode.com/svn/trunk@711 59da171f-624f-0410-aa54-27559c288bec