Update issue #353 Status: Fixed
The outer-hash fix was done with r711.
In r711 there is also an option to run Duplicati.CommandLine with the "verify" command.
Finally r711 also has an option to create a verification file on the backend,
using the --create-verification-file option.

This commit adds a small python script that can be used to verify the files
on the remote backend without using the backup password.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@712 59da171f-624f-0410-aa54-27559c288bec