Update issue #378 Status: Fixed
I have now added a module called http-options which can be used to
configure http options, notably the 100-Expect.

I am unsure if it will break existing installations if I disable
it by default, so it is still enabled but can be disabled by the
switch --disable-expect100-continue.

You can set this option in the GUI under "Advanced options".

git-svn-id: https://duplicati.googlecode.com/svn/trunk@713 59da171f-624f-0410-aa54-27559c288bec