Update issue #376 Status: Fixed
There is now a Google Docs backend.

The API turned out to be broken (or at least hard to use correctly) for uploading files,
so I eventually wrote a custom uploader. The custom uploader can also upload files larger
than 10 MB.

I have tested this with my usual test setup, but let me know if there are any issues with it.
A preview release will be put on the download page asap.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@835 59da171f-624f-0410-aa54-27559c288bec