Update issue #458 Status: Fixed
There is now a delete transaction file being written, which should prevent cases where a large number of orphan files are detected.
If a large number of orphan files are detected, this should now be considered a real error.

I chose not to put in the cleanup button, because it would take too long for me to fit it in there, and I would rather focus on getting 2.0 ready, which will have a new UI anyway.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@957 59da171f-624f-0410-aa54-27559c288bec