Update issue #509 Status: Fixed
The problem is that the the file-based backend returns "Directory not found" when asking for the not-yet-mounted folder.
Duplicati assumed that this error means that the folder will be created later, and thus returned an empty list,
as the folder would be empty when created. This means that the retry logic was not used. 

Backing up to an empty folder will then result in a full backup being made.

But once the files started uploading, the folder is there and the folder is not created, but it is still a full backup.

This commit changes the logic a bit, so the list operation will not return until successful and a missing folder will be created during the list operation.
This ensures that the backup does not start before there has been a successful listing of the folder. If it fails, there is the normal retry delay, which should be enough to ensure that the folder is re-mounted. There may be warnings in the log, that the auto create failed. I have removed the logic that will auto create the folder during upload, as it should no longer be possible to get to the upload stage without verifying that the folder exists.

I have also changed the logic of the --disable-autocreate-folder option, so default value is "false" if it is a backup, and "true" for any other operation, so simply issuing a "list" or "cleanup" will not create folders (it did not do that before either).


git-svn-id: https://duplicati.googlecode.com/svn/trunk@1259 59da171f-624f-0410-aa54-27559c288bec