Update issue #627 Status: Fixed
I changed the implementation so it uses Mono.Data.Sqlite if that is present,
and then fallback to System.Data.SQLite (calling libsqlite).

This fixes the crash on my setup, and I was able to complete a SSH based backup without issues on Ubuntu 12.04.

To prevent compatibility issues, it is possible to revert to the old way of just using
System.Data.SQLite by setting the environment variable "DISABLE_MONO_DATA_SQLITE".

git-svn-id: https://duplicati.googlecode.com/svn/branches/1.3.2@1302 59da171f-624f-0410-aa54-27559c288bec