Update issue #766
Refactored the patch from issue #766
Moved reading into a function in the utility module.
Moved parsing of file data into a function.

Added parameters to the command line client, such that the CLI may have options.
Implemented these options in the help output, so the --parameters-file can be viewed from the help.

git-svn-id: https://duplicati.googlecode.com/svn/trunk@1511 59da171f-624f-0410-aa54-27559c288bec