XML Reader Helpers
Wrote a function to load a XML file. We always copy the XML to the
tempfolder since it might be used by another process when we try to load
the file.

The other function will retrieve the associated active VHD of the
configuration.

The next step would be to copy certain configurations settings from the
backup XML to implement in the previously created Hyper-V machine.