Merging AVHD's and VHD's
Implemented a feature to merge all VHD's in one. We could present this
as an option the the end-user.

It would also make up for a cleaner backup.

The XML still needs to be edited after the merge so that the config will
point to the parent VHD. Also the snapshots should be removed, these can
still be viewed in the Hyper-V Manager.