Rewrote the m_sourcepaths as a List
Tried to stay clear of existing code but I eventually changed this to be
a List instead of an array. It's much more flexible. There really is no
performance gain with Arrays unless when we know that data is fixed
length and we want to micro-optimise.

Removed lot's of comments since all documentation can come here now.

Note: HyperV vss-writer; are already implicitly included in the
vss-writers vss-writer will disable all other vss-writers.