This is a two fold "fix" to duplicati/duplicati#2715
Changes in settings.html adds the option to select "0 seconds", which wasn't previously possible.
Changes in SystemSettingsController.js adds logic in the controller that will check if the duration value and duration multipliers are set. If it finds that they are not it will set them to "0" and "s" respectively (instead of leaving them blank)

Issue comment called for a check box to enable/disable startup delay, but just defaulting to 0 seconds is simpler and self explanitory.