Fix Travis build issues by limiting concurrent processes.
It seems that msbuild has issues when building with multiple concurrent
processes.  Once we restricted the build by providing the /maxcpucount:1
option and using a single job in the Travis build, we:

1. Obtained 5 consecutive successful builds.
2. We then removed the /maxcpucount:1 option and immediately encountered
   a failed build.
3. Adding back the /maxcpucount:1 option, we then obtained another 10
   consecutive successful builds.

Prior to these changes, 6 out of 9 builds failed (with msbuild hanging)
while experimenting with various options along with numerous failures in
previous pull requests, etc.

After limiting Travis to a single job, the build times remain
approximately the same as before (22-24 minutes).

Note that similar to the AppVeyor configuration, we skip the "BulkData"
tests.

See pull request #3171 for experimental details.