Convert regex pattern to wildcard in default excludes group.
This provides a workaround for the poor performance of the default
excludes group exhibited by regex filters.  While this wildcard pattern
will potentially exclude more files than the previous regex pattern, in
practice this should not often be the case.

This does not resolve the performance issue of the regex filters.
However, I think in general we should try to avoid expensive
application-specific filters since these will affect the performance for
all users that use the default excludes group, whether the application
is present or not.

This concerns issue #3395.