Ensure removal of VolumeWriterBase temporary files.
Each VolumeWriterBase contains a TempFile, which deletes a file upon
disposal.  If we do not dispose of this instance, temporary files may be
left behind.

Some of these instances could not be put inside a using block, as the
variables are being redefined inside a loop (a using block marks the
object as readonly).

This concerns issue #3652.