Revert "Remove unused YouMayGetMoreHelpHere method."
This reverts commit 2c9fabd2edcac8727bdbbba72fd2d1cf07aae6da.

This is a placeholder message that is intended to be used with the code
for each error and log message. The idea is that the commandline will
use the code to provide a link to the forum, searching for the code,
such that it is easy to maintain a list of help links.