Fix 'Stop after upload' on backup
'Stop after upload' currently does not stop a running backup but lets the backup fully complete. This leaves the only alternative to force the job to stop using 'Stop now'.

Now when a backup job is running and 'Stop after upload' is chosen the backup job will complete the current file and then cleanly stop the job.

Should 'Stop after upload' be renamed to something like 'Stop after current file' ?

note: 'Stop now' abruptly kills the thread and causes the backup to be left in a bad state. Perhaps this should not be a readily available option to users? Or the 'Stop now' could be changed to cleanly abort the current file so the backup is left in a good state.