Avoid possible invalid DateTimeOffset values.
When creating a new backup, a value of DateTime.MinValue can be provided
to the NormalizeDateTimeToEpochSeconds method.  If DateTimeOffset is
used, it will convert the provided DateTime to the UTC equivalent, which
may result in a value outside of the allowable bounds.

This reverts the changes from pull request #3919.
This fixes issue #3955.